Test struct in contoller. Move structs to different file. Update gitignore
