+++ .gitignore
@@ -15,6 +15,8 @@ node_modules/
 .cache/
 dist/
 
+/back/vendor
+/back/mongodb
 /back/back.exe
 /back/back
 /back/main.exe
